+++ includes/class-rest-capabilities.php
@@ -486,6 +486,7 @@ class Capabilities extends Base {
 			'toegangscontrole'   => 'Toegangscontrole',
 			'manage_clothing'    => 'Kledingbeheer',
 			'ledenadministratie' => 'Ledenadministratie',
+			'sponsorbeheer'      => 'Sponsorbeheer',
 			'vrijwilligers'      => 'Vrijwilligersbeheer',
 			'rondo_iva_approve'  => 'IVA goedkeuren',
 		];
@@ -550,7 +551,7 @@ class Capabilities extends Base {
 			);
 		}
 
-		$allowed_caps  = [ 'fairplay', 'vog', 'financieel', 'financieel_read', 'toegangscontrole', 'manage_clothing', 'ledenadministratie', 'vrijwilligers', 'rondo_iva_approve' ];
+		$allowed_caps  = [ 'fairplay', 'vog', 'financieel', 'financieel_read', 'toegangscontrole', 'manage_clothing', 'ledenadministratie', 'sponsorbeheer', 'vrijwilligers', 'rondo_iva_approve' ];
 		$valid_slugs   = array_keys( \Rondo\Core\UserRoles::get_all_roles() );
 		$valid_slugs[] = 'administrator';