+++ includes/class-rest-user-settings.php
@@ -1028,6 +1028,7 @@ class UserSettings extends Base {
 			|| current_user_can( 'toegangscontrole' )
 			|| current_user_can( 'manage_clothing' )
 			|| current_user_can( 'ledenadministratie' )
+			|| current_user_can( 'sponsorbeheer' )
 			|| current_user_can( 'vrijwilligers' );
 
 		$person_id           = (int) get_user_meta( $user_id, 'rondo_linked_person_id', true );
@@ -1061,6 +1062,7 @@ class UserSettings extends Base {
 			'has_extra_roles'               => $has_extra_roles,
 			'is_kader'                      => $is_kader,
 			'can_edit_people'               => \Rondo\Core\AccessControl::can_edit_people(),
+			'can_manage_sponsors'           => \Rondo\Core\AccessControl::can_manage_sponsors(),
 			'can_access_fairplay'           => current_user_can( 'fairplay' ),
 			'can_access_vog'                => current_user_can( 'vog' ),
 			'can_access_financieel'         => \Rondo\Core\UserRoles::can_view_finances(),